+++ lib/ui/profile/ProfileInfoPreviewPage.dart
@@ -133,7 +133,7 @@ class ProfileInfoPreviewPageState extends State<ProfileInfoPreviewPage> {
       pronunciationAudioData: widget.pronunciationAudioData,
     );*/
     NotificationService().notify(ProfileHomePanel.notifySelectContent, [
-      ProfileContent.share,
+      ProfileContentType.share,
       <String, dynamic>{
         ProfileInfoSharePage.profileResultKey : ProfileInfoLoadResult(
           profile: _profile,